requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
flake8==3.4.1
flake8-import-order==0.13
future==0.16.0
idna==2.6
mccabe==0.6.1
oauthlib==2.0.6
pycodestyle==2.3.1
pyflakes==1.5.0
python-twitter==3.3
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
